back/discount-django/requirements.txt
asgiref==3.3.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
et-xmlfile==1.0.1
idna==3.3
itypes==1.2.0
Jinja2==3.0.3
MarkupPy==1.14
MarkupSafe==2.0.1
oauthlib==3.1.1
odfpy==1.4.1
openpyxl==3.0.7
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
tablib==3.0.0
uritemplate==4.1.1
urllib3==1.26.7
xlrd==2.0.1
xlwt==1.3.0
